requirements.txt
asgiref==3.11.0
attrs==25.4.0
click==8.3.1
colorama==0.4.6
croniter==6.0.0
Django==6.0.1
django-cors-headers==4.9.0
django-redis==6.0.0
django-rq==3.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
rq==2.6.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.11.0
